+++ RELEASES.md
@@ -12,6 +12,7 @@ This release will be the final version that uses Poetry for package and dependen
 - Added bars update support for `Portfolio` PnLs (#2239), thanks @faysou
 - Added optional `params` for `Strategy` order management methods (symmetry with `Actor` data methods) (#2251), thanks @faysou
 - Added heartbeats for Betfair clients to keep streams alive (more robust when initial subscription delays)
+- Added market start time range filtering for `BetfairInstrumentProviderConfig`
 - Improved Betfair clients configuration
 
 ### Breaking Changes